Beginnings of SudokuX extension.
Sudoku-X is basically classic Sudoku, with the additional constraints
that the diagonals (i.e. the "X" from the name) must also be unique.
As such, it naturally makes sense that we should try to reuse as much
of the classic Sudoku stuff as possible.

The approach started here tries to do this by piggybacking off the
standard Sudoku puzzle. This is done by introducing the concept of
"prerequisites" (i.e. dependencies) which must be available from the
ExtensionManager (to be looked up at runtime) for the extension to
be able to be used. So, even if an extension is registered, it may not
be usable (and hence should be considered non-existent).

The current implementation only includes some initial sketches of some
parts of this system, and doesn't compile yet.